+++ analyzer/src/test/java/nl/basjes/parse/useragent/analyze/TestBuilder.java
@@ -228,9 +228,18 @@ public class TestBuilder {
         UserAgentAnalyzer uaa = builder.build();
         assertNotNull(uaa, "We should get a first instance from a single builder.");
 
+        // Try to load something that does not yield any files.
+        // Optional --> error message and continue
+        uaa.loadResources("Bad resource string that is optional should only give a warning", true, true);
+
+        // NOT Optional --> fail with exception
+        assertThrows(InvalidParserConfigurationException.class, () ->
+            uaa.loadResources("Bad resource string that is NOT optional should fail hard", true, false)
+        );
+
         uaa.initializeMatchers();
         assertThrows(IllegalStateException.class, () ->
-            uaa.loadResources("Something extra")
+            uaa.loadResources("Cannot load resources after initialization")
         );
     }